requirements.txt
# LangChain 1.0 and LangGraph
langchain>=1.0.0
langchain-core>=1.0.0
langchain-openai>=1.0.0
langchain-anthropic>=0.3.0
langgraph>=0.2.0

# MCP (Model Context Protocol)
mcp>=1.0.0

# API clients
amadeus>=8.0.0
requests>=2.31.0
httpx>=0.27.0

# Environment and utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Date handling
python-dateutil>=2.8.2

# CLI and user interaction
rich>=13.0.0
prompt-toolkit>=3.0.0
